Fix narrowing conversions in ccsd_driver, 2-, 4-index transform code

+++ tammx/two_four_index.cc
@@ -1,6 +1,6 @@
 #include "hartree_fock.h"
 
-std::tuple<Tensor4D> two_four_index_transform(const int ndocc, const int nao, const int freeze_core, const int freeze_virtual, const Matrix &C, Matrix &F, libint2::BasisSet &shells){
+std::tuple<Tensor4D> two_four_index_transform(const size_t ndocc, const size_t nao, const size_t freeze_core, const size_t freeze_virtual, const Matrix &C, Matrix &F, libint2::BasisSet &shells){
 
   using libint2::Atom;
   using libint2::Shell;
